# In `add_months()`, conditionally attempt to add months